Add files via upload
A Netflix clone made using only HTML and CSS for the frontend is a web application that closely mimics the design and user interface of the popular streaming platform Netflix. This clone provides a way for users to browse and watch a collection of movies and TV shows just like they would on the actual platform, but without any backend functionality.